requirements.txt
deepface
opencv-python
pandas
mediapipe
pillow
numpy
python-dotenv
